requirements.txt
fastapi
uvicorn
flask
python-dotenv
requests
msal